realloc_08: Move functions our of cli to portfolio.py
